+++ crates/language-v2/definition/src/model/utils/parser_options.rs
@@ -10,7 +10,10 @@ pub struct ParserOptions {
     pub inline: bool,
 
     /// Whether to make the parsing rule public, generating a parser
-    /// that can be used from outside the module.    
+    /// that can be used from outside the module.
+    ///
+    /// TODO(v2): This argument is currently used for testing different non-terminals, once we start testing everything at the `SourceUnit`
+    /// level we may want to get rid of it
     pub public: bool,
 
     /// Verbatim code to insert into the parser rule.
