+++ crates/solidity-v2/outputs/cargo/parser/Cargo.toml
@@ -19,6 +19,8 @@ keywords = ["parser"]
 categories = ["compilers", "parsing", "parser-implementations"]
 
 [features]
+# TODO(v2): Once we split the testing for V2, we should consider moving the testing utilities into a separate
+# crate, and have these dependencies only in there
 __private_testing_utils = ["anyhow", "infra_utils"]
 
 [build-dependencies]
