+++ crates/solidity-v2/outputs/cargo/parser/src/lexer/definition.rs
@@ -130,26 +130,3 @@ impl Lexer<'_> {
         first_part
     }
 }
-
-/// Iterate over the lexemes and their offsets
-///
-/// TODO(v2): This iterator skips all trivia, we'll want to include it in
-/// future versions
-impl Iterator for Lexer<'_> {
-    type Item = Result<(usize, LexemeKind, usize), ()>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some(lexeme) = self.next_lexeme() {
-            match lexeme.kind {
-                LexemeKind::Whitespace
-                | LexemeKind::EndOfLine
-                | LexemeKind::SingleLineComment
-                | LexemeKind::MultiLineComment
-                | LexemeKind::SingleLineNatSpecComment
-                | LexemeKind::MultiLineNatSpecComment => {}
-                _ => return Some(Ok((lexeme.range.start, lexeme.kind, lexeme.range.end))),
-            }
-        }
-        None
-    }
-}
