+++ crates/solidity-v2/outputs/cargo/parser/src/parser/grammar.lalrpop.jinja2
@@ -1,6 +1,6 @@
 use crate::lexer::LexemeKind;
 use slang_solidity_v2_cst::structured_cst::nodes::*;
-use crate::parser::nodes::*;
+use crate::parser::parser_helpers::*;
 
 grammar<'source>(source: &'source str);
