+++ crates/solidity/testing/snapshots/cst_output/SourceUnit/unrecognized_lexeme_after_contract/input.sol
@@ -1,0 +1,1 @@
+contract X # { }
